Add Jest tests for listing helpers
